doc: update process section fix: fixed some compiler warnings

+++ kernel/sched/sched.c
@@ -64,5 +64,5 @@ u64 switch_context(void) {
      * Return the correct value in order to make eret_to_thread work correctly
      * in main.c
      */
-    return target_ctx;
+    return (u64)target_ctx;
 }
